fix: time ratio using floating point

+++ src/components/virtual_time/vtime.rs
@@ -5,7 +5,7 @@ use crate::qemu_api::qemu_plugin_get_snapshot_cpu_clock;
 pub struct VirtualTimeContext {
     last_real_time: u64,
     advanced_vclock: i64,
-    time_scaling_factor: u64,
+    time_scaling_factor: f64,
 }
 
 impl VirtualTimeContext {
@@ -13,12 +13,12 @@ impl VirtualTimeContext {
         Self {
             last_real_time: 0,
             advanced_vclock: 0,
-            time_scaling_factor: param::INIT_HOST_TIME_SCALE as u64,
+            time_scaling_factor: param::INIT_HOST_TIME_SCALE as f64,
         }
     }
 
-    pub fn update_scaling_factor(&mut self, scaling_factor: u64) {
-        if scaling_factor == 0 {
+    pub fn update_scaling_factor(&mut self, scaling_factor: f64) {
+        if scaling_factor == 0.0 {
             panic!("The scaling factor should not be zero.");
         }
         self.time_scaling_factor = scaling_factor;
@@ -38,7 +38,7 @@ impl VirtualTimeContext {
             if qemu_plugin_cpu_is_tick_enabled() && self.last_real_time != 0 {
                 // 3.2 if the maximum is zero, we use the difference of the real time.
                 let advanced_vtime =
-                    ((real_time - self.last_real_time) / self.time_scaling_factor) as i64;
+                    ((real_time - self.last_real_time) as f64 / self.time_scaling_factor) as i64;
 
                 // 3.3 update the advanced vclock
                 self.advanced_vclock += advanced_vtime;
